exercise_05/todo-app/backend/requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
mysql-connector==2.2.9
pymysql==0.9.3
